backend/foodgram/requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
django==2.2.19
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.0.4
idna==3.3
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
oauthlib==3.2.0
pandas==1.3.5
Pillow==9.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.1 